Added test points to BMS chip.
Tested BMS functionality and it is not working. It will prevent the battery from being charged at a high voltage, but it will not allow the battery to discharge under normal conditions. Interestingly, attaching 5V to VBUS or VCC while the battery is installed, then removing it, will "kick start" the BMS and allow discharging from the battery.